polygon1.c: Fix possible case where we could return an invalid polygon upon allocation failure
This was a bug present in the poly_Boolean() routine. We need to set the output
polygon to NULL if we are going to return early on memory failure. Comment this,
and add a similar note to poly_Boolean_free(), where we already set the return
value (and hence there was no bug).