polygon1.c: Fix possible case where we could return an invalid polygon upon allocation failure
This was a bug present in the poly_Boolean() routine. We need to set the output
polygon to NULL if we are going to return early on memory failure. This was
noticed during general review after observing problems due to our copy routines
incorrectly returning FALSE when copping a NULL polygon.