Fixup square thermal polygons to use frac_circle2
The main adjustment is that callers need to add their own last vertex
in order to avoid an extraneous edge marked as being round.

XXX: Might be neater to fix this by having frac_circle just not set
the round flag on the last vertex added... we can extend the checking
code to re-use vertices if the next edge added is to be round.